owl_description/requirements.txt
clip==1.0
moveit_commander==1.1.12
numpy==1.23.4
open3d==0.13.0
opencv_contrib_python==4.7.0.72
opencv_python==4.7.0.72
Pillow==10.0.0
pyrealsense2==2.53.1.4623
python_pcl==0.3
PyYAML==6.0
PyYAML==6.0.1
rospy==1.16.0
scipy==1.11.1
sensor_msgs==1.13.1
tf==1.13.2
tf2_geometry_msgs==0.7.6
tf2_ros==0.7.6
torch==2.0.1
transformers==4.31.0
